Implement "did you mean?" feature
Adds a new (true by default) `--suggestions` flag to the CLI tool, and a
`suggestions` (true by default) GitHub Action flag. The flag exists
because this feature does require the tool to be run inside of a Git
repository (though not necessarily at the root).

The feature is implemented by considering all files in the repository
(from git ls-files) as link candidates, and finding the relative path
link with the shortest edit-distance from the source file.